+++ R/data.R
@@ -1,11 +1,23 @@
-#' Dataset for demonstration purposes
+#' Critical Power Dataset
 #'
 #' A dataset containing 5 time-to-exhaustion trials for demonstrating
-#' how to use the package.
+#' how to use the `critical_power()` function.
 #'
 #' @format A data frame with 5 rows and 2 variables:
 #' \describe{
 #'   \item{PO}{power output values, in watts.}
 #'   \item{TTE}{time-to-exhaustion, in seconds, from the corresponding power output.}
 #' }
-"demo_data"
+"demo_critical_power"
+
+#' Critical Speed Dataset
+#'
+#' A dataset containing 5 time-to-exhaustion trials for demonstrating
+#' how to use the `critical_speed()` function.
+#'
+#' @format A data frame with 3 rows and 2 variables:
+#' \describe{
+#'   \item{Distance}{distance values, in meters}
+#'   \item{TTE}{time-to-exhaustion, in seconds, from the corresponding power output.}
+#' }
+"demo_critical_speed"