+++ R/utils.R
@@ -81,9 +81,9 @@ get_combinations <- function(
 #' @export
 #'
 #' @examples
-#' combinations <- get_combinations(demo_data, "PO", "TTE")
+#' combinations <- get_combinations(demo_critical_power, "PO", "TTE")
 #'
-#' results_index <- get_indexes(demo_data, "PO", "TTE", combinations)
+#' results_index <- get_indexes(demo_critical_power, "PO", "TTE", combinations)
 #'
 #' results_index
 get_indexes <- function(